brent_spence_bridge/backend/requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==3.0.10
mysql-connector-python==8.0.26
pandas==2.3.1
python-dotenv==0.19.0
torch==2.2.0
scikit-learn==1.6.1
numpy<2.0.0
matplotlib>=3.9.0